tasks/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-rest-auth==5.0.1
Django==4.1.11
django-allauth==0.56.1
django-cors-headers==4.2.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
oauthlib==3.2.2
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
sqlparse==0.4.4
